feat:display the error message according to the backend
